Restructured project: smallers files as opposed to one big file
