+++ flightplugin/Logger.cpp
@@ -1,0 +1,32 @@
+#include "Logger.h"
+
+using namespace sp;
+
+void Logger::log(std::string s)
+{
+	cvarManager->log(s);
+}
+void Logger::log(Vector v)
+{
+	cvarManager->log(sp::vector_to_string(v));
+}
+void Logger::log(Rotator r)
+{
+	cvarManager->log(sp::rot_to_string(r));
+}
+void Logger::log(float f)
+{
+	cvarManager->log(sp::to_string(f));
+}
+void Logger::log(bool b)
+{
+	cvarManager->log(to_string(b));
+}
+void Logger::log(int i)
+{
+	cvarManager->log(to_string(i));
+}
+void Logger::log(unsigned long l)
+{
+	cvarManager->log(to_string(l));
+}
