+++ flightplugin/Painter.cpp
@@ -1,0 +1,375 @@
+#include "bakkesmod\wrappers\GameEvent\TutorialWrapper.h"
+#include "bakkesmod\wrappers\GameObject\CarWrapper.h"
+#include "bakkesmod\wrappers\GameObject\CarComponent\VehicleSimWrapper.h"
+#include "bakkesmod\wrappers\arraywrapper.h"
+#include "bakkesmod\wrappers\wrapperstructs.h"
+#include "Painter.h"
+#include "CarManager.h"
+#include "Arrow.h"
+#include "HelperFunctions.h"
+#include "flightplugin.h"
+
+using namespace sp;
+
+void Painter::initDrawables()
+{
+	gameWrapper->RegisterDrawable([this](CanvasWrapper cw) {
+		if (!gameWrapper->IsInGame()) {
+			return;
+		}
+		if (cvarManager->getCvar("showHUD").getBoolValue())
+		{
+			this->drawPanels(cw);
+		}
+		});
+}
+void Painter::drawPanels(CanvasWrapper cw)
+{
+	auto players = gameWrapper->GetGameEventAsServer().GetCars();
+	if (players.Count() > 0)
+	{
+		CarWrapper car = players.Get(0);
+		RBState rbstate = car.GetRBState();
+		if (cvarManager->getCvar("showCarDerivedInfo").getBoolValue())
+		{
+			drawCarDerivedInfo(cw, car, 20, 55);
+		}
+		if (cvarManager->getCvar("showYaw").getBoolValue())
+		{
+			drawYawPlane(cw, car, 20, 700, 1);
+		}
+	}
+}
+
+void Painter::drawStringAt(CanvasWrapper cw, std::string text, int x, int y, Color col)
+{
+	cw.SetPosition({ x, y });
+	cw.SetColor(col.r, col.g, col.b, col.a);
+	cw.DrawString(text);
+}
+
+void Painter::drawStringAt(CanvasWrapper cw, std::string text, Vector2_ loc, Color col)
+{
+	drawStringAt(cw, text, loc.X, loc.Y, col);
+}
+
+void Painter::drawLine(CanvasWrapper cw, Vector2_ v1, Vector2_ v2)
+{
+	cw.DrawLine({ v1.X, v1.Y }, { v2.X, v2.Y });
+}
+
+void Painter::drawRBStatePanel(CanvasWrapper cw, std::string title, RBState rbstate, int x, int y, bool recording)
+{
+	Vector loc = rbstate.Location;
+	Vector lin = rbstate.LinearVelocity;
+	Quat quat = rbstate.Quaternion;
+	Vector ang = rbstate.AngularVelocity;
+	int marginLeft = 30;
+	int marginTop = 20;
+	int nameSpacing = 100;
+	int vecSpacing = 80;
+	int quatSpacing = 120;
+	int lineSpacing = 30;
+	int width = 620;
+	int height = 210;
+	cw.SetPosition({ x, y });
+	cw.SetColor(COLOR_PANEL);
+	cw.FillBox({ width, height });
+	cw.SetColor(COLOR_TEXT);
+
+	this->drawStringAt(cw, title, x + marginLeft, y + marginTop);
+	int currentLine = marginTop + 50;
+	this->drawStringAt(cw, "Location", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(loc.X, 2), x + marginLeft + nameSpacing, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(loc.Y, 2), x + marginLeft + nameSpacing + vecSpacing, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(loc.Z, 2), x + marginLeft + nameSpacing + vecSpacing * 2, y + currentLine);
+	currentLine += lineSpacing;
+	this->drawStringAt(cw, "Lin. Velocity", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(lin.X, 2), x + marginLeft + nameSpacing, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(lin.Y, 2), x + marginLeft + nameSpacing + vecSpacing, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(lin.Z, 2), x + marginLeft + nameSpacing + vecSpacing * 2, y + currentLine);
+	currentLine += lineSpacing;
+	this->drawStringAt(cw, "Rotation", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string_scientific(quat.W), x + marginLeft + nameSpacing, y + currentLine);
+	this->drawStringAt(cw, sp::to_string_scientific(quat.X), x + marginLeft + nameSpacing + quatSpacing, y + currentLine);
+	this->drawStringAt(cw, sp::to_string_scientific(quat.Y), x + marginLeft + nameSpacing + quatSpacing * 2, y + currentLine);
+	this->drawStringAt(cw, sp::to_string_scientific(quat.Z), x + marginLeft + nameSpacing + quatSpacing * 3, y + currentLine);
+	currentLine += lineSpacing;
+	this->drawStringAt(cw, "Ang. Velocity", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ang.X, 4), x + marginLeft + nameSpacing, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ang.Y, 4), x + marginLeft + nameSpacing + vecSpacing, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ang.Z, 4), x + marginLeft + nameSpacing + vecSpacing * 2, y + currentLine);
+}
+
+void Painter::drawCarDerivedInfo(CanvasWrapper cw, CarWrapper car, int x, int y)
+{
+	RBState rbstate = car.GetRBState();
+	Vector loc = rbstate.Location;
+	Vector lin = rbstate.LinearVelocity; //this lin.magnitude = speed relative to world
+	Quat quat = rbstate.Quaternion;
+	Vector ang = rbstate.AngularVelocity;
+	auto horVel = Vector(lin.X, lin.Y, 0);
+	Vector up = quatToUp(quat);
+	Vector right = quatToRight(quat);
+	Vector fwd = quatToFwd(quat);
+	auto linLocalFwd = Vector::dot(lin, fwd);
+	auto linLocalRight = Vector::dot(lin, right);
+	auto linLocalUp = Vector::dot(lin, up);
+	Vector linLocal = Vector(linLocalFwd, linLocalRight, linLocalUp);
+	auto lonSpeed = Vector::dot(horVel, fwd); //speed relative to cars forward
+	auto latSpeed = Vector::dot(horVel, right);
+	//	auto vsim = car.GetVehicleSim();
+	//	auto wheels = vsim.GetWheels();
+	int marginLeft = 10;
+	int marginTop = 20;
+	int nameSpacing = 100;
+	int vecSpacing = 120;
+	int quatSpacing = 120;
+	int lineSpacing = 30;
+	cw.SetPosition({ x, y });
+	cw.SetColor(COLOR_PANEL);
+	cw.FillBox({ 300, 375 });
+	cw.SetColor(COLOR_TEXT);
+	cw.SetColor(205, 155, 15, 255);
+	this->drawStringAt(cw, "Flight Plugin", x + marginLeft, y + marginTop);
+	int currentLine = marginTop + 20;
+	cw.SetColor(255, 255, 255, 255);
+	this->drawStringAt(cw, "Speed", x + marginLeft, y + currentLine); //speed relative to cars forward
+	this->drawStringAt(cw, sp::to_string(lonSpeed, 4), x + marginLeft + nameSpacing, y + currentLine);
+	currentLine += lineSpacing;
+	this->drawStringAt(cw, "Lat. speed", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(latSpeed, 4), x + marginLeft + nameSpacing, y + currentLine);
+	currentLine += lineSpacing;
+	this->drawStringAt(cw, "Linear speed", x + marginLeft, y + currentLine); //this lin.magnitude = speed relative to world
+	this->drawStringAt(cw, sp::to_string(lin.magnitude(), 4), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+	this->drawStringAt(cw, "up", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::vector_to_string(up, 2), x + marginLeft + nameSpacing, y + currentLine);
+	currentLine += lineSpacing;
+	this->drawStringAt(cw, "right", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::vector_to_string(right, 2), x + marginLeft + nameSpacing, y + currentLine);
+	currentLine += lineSpacing;
+	this->drawStringAt(cw, "fwd", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::vector_to_string(fwd, 2), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+	this->drawStringAt(cw, "quat", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::quat_to_string(quat, 2), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+	this->drawStringAt(cw, "ang", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::vector_to_string(ang, 2), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+	this->drawStringAt(cw, "linLocalUp", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(linLocalUp, 2), x + marginLeft + nameSpacing, y + currentLine);
+	currentLine += lineSpacing;
+	this->drawStringAt(cw, "linLocalRight", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(linLocalRight, 2), x + marginLeft + nameSpacing, y + currentLine);
+	currentLine += lineSpacing;
+	this->drawStringAt(cw, "linLocalFwd", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(linLocalFwd, 2), x + marginLeft + nameSpacing, y + currentLine);
+
+}
+
+void Painter::drawVector(CanvasWrapper canvas, Vector vec)
+{
+	if (gameWrapper->IsInGame()) // TODO: readd flight_enabled if necessary
+	{
+		ServerWrapper game = gameWrapper->GetGameEventAsServer();
+
+		if (game.IsNull())
+			return;
+		ArrayWrapper<CarWrapper> cars = game.GetCars();
+
+		if (cars.Count() > 0) {
+			CarWrapper car = cars.Get(0);
+			if (car.IsNull())
+				return;
+
+			if (arrow.size() == 0) { // arrow on car
+				arrow = CarManager::getArrowPoints(static_cast<CARBODY>(*carType), *gameWrapper);
+			}
+			canvas.SetColor(255, 255, 0, 200);
+
+			Vector v = car.GetLocation();
+			Rotator r = car.GetRotation();
+
+			double dPitch = (double)r.Pitch / 32764.0 * 3.14159;
+			double dYaw = (double)r.Yaw / 32764.0 * 3.14159;
+			double dRoll = (double)r.Roll / 32764.0 * 3.14159;
+
+			Vector2 carLocation2D = canvas.Project(v);
+			Vector2 arrow2D[2];
+			for (int i = 0; i < 2; i++) {
+				arrow2D[i] = canvas.Project(Rotate(arrow[i], dRoll, -dYaw, dPitch) + v); 
+			}
+
+			canvas.DrawLine(arrow2D[0], arrow2D[1]); // Actually draws the arrow from point 0 to 1.
+
+			canvas.SetPosition(carLocation2D.minus({ 10,10 }));
+			canvas.FillBox({ 20, 20 });
+			return;
+		}
+	}
+}
+
+Vector Painter::Rotate(Vector aVec, double roll, double yaw, double pitch)
+{
+
+	float sx = sin(roll);
+	float cx = cos(roll);
+	float sy = sin(yaw);
+	float cy = cos(yaw);
+	float sz = sin(pitch);
+	float cz = cos(pitch);
+
+	aVec = Vector(aVec.X, aVec.Y * cx - aVec.Z * sx, aVec.Y * sx + aVec.Z * cx);  //2  roll?
+	aVec = Vector(aVec.X * cz - aVec.Y * sz, aVec.X * sz + aVec.Y * cz, aVec.Z); //1   pitch?
+	aVec = Vector(aVec.X * cy + aVec.Z * sy, aVec.Y, -aVec.X * sy + aVec.Z * cy);  //3  yaw?
+
+	// ugly fix to change coordinates to Unreal's axes
+	float tmp = aVec.Z;
+	aVec.Z = aVec.Y;
+	aVec.Y = tmp;
+	return aVec;
+}
+
+void Painter::drawYawPlane(CanvasWrapper cw, CarWrapper car, int x, int y, float scale)
+{
+	RBState rbstate = car.GetRBState();
+	Vector loc = rbstate.Location;
+	Vector lin = rbstate.LinearVelocity;
+	Quat quat = rbstate.Quaternion;
+	Vector ang = rbstate.AngularVelocity;
+	Vector up = quatToUp(quat);
+	Vector right = quatToRight(quat);
+	Vector fwd = quatToFwd(quat);
+	Vector fwdHor(fwd.X, fwd.Y, 0);
+	fwdHor.normalize();
+	Vector velHor(lin.X, lin.Y, 0);
+	if (velHor.magnitude() > 0)
+	{
+		float fraction = (velHor.magnitude() / 2300) * 100;
+		velHor.normalize();
+		velHor = velHor * fraction; // scale vector to range 0-100 to fit it in the panel
+	}
+	Vector2_ velArrow(velHor.Y, -velHor.X);
+	Vector2_ offset(x, y);
+	int width = 200;
+	int height = 200;
+	cw.SetPosition({ x, y });
+	cw.SetColor(COLOR_PANEL);
+	cw.FillBox({ (int)(width * scale), (int)(height * scale) });
+	cw.SetColor(COLOR_TEXT);
+	Vector2_ center(100, 100);
+	Vector2_ axisVer(0, -80);
+	Vector2_ axisHor(80, 0);
+	Vector2_ orientArrow(0, -90);
+	Vector2_ angleLoc(10, 10);
+	Vector2_ axisVerLabelLoc(-15, -85);
+	Vector2_ axisHorLabelLoc(75, 5);
+	drawLine(cw, offset + center * scale, offset + (center + axisVer) * scale);	// x-axis
+	drawLine(cw, offset + center * scale, offset + (center + axisHor) * scale);	// y-axis
+	float cosTheta = Vector::dot(fwdHor, Vector(1, 0, 0));
+	float theta = acos(cosTheta);
+	if ((Vector::cross(Vector(1, 0, 0), fwdHor)).Z > 0) { theta = -theta; }
+	Vector2_ frontLeft(-42, -73);	// approximate shape/size of octane
+	Vector2_ frontRight(42, -73);
+	Vector2_ backLeft(-42, 45);
+	Vector2_ backRight(42, 45);
+	drawLine(cw, offset + (center + frontLeft.rotate(theta)) * scale, offset + (center + frontRight.rotate(theta)) * scale);	// front and back of car
+	drawLine(cw, offset + (center + backLeft.rotate(theta)) * scale, offset + (center + backRight.rotate(theta)) * scale);
+	drawLine(cw, offset + (center + frontLeft.rotate(theta)) * scale, offset + (center + backLeft.rotate(theta)) * scale);
+	drawLine(cw, offset + (center + frontRight.rotate(theta)) * scale, offset + (center + backRight.rotate(theta)) * scale);
+	cw.SetColor(255, 255, 255, 128);	// fwd arrow
+	drawLine(cw, offset + center * scale, offset + (center + orientArrow.rotate(theta)) * scale);
+	cw.SetColor(0, 255, 0, 255);	// velocity arrow
+	drawLine(cw, offset + center * scale, offset + (center + velArrow) * scale);
+	this->drawStringAt(cw, sp::to_string(theta), offset + angleLoc * scale);
+	this->drawStringAt(cw, "x", offset + (center + axisVerLabelLoc) * scale);
+	this->drawStringAt(cw, "y", offset + (center + axisHorLabelLoc) * scale);
+}
+void Painter::drawInputPanel(CanvasWrapper cw, CarWrapper car, int x, int y)
+{
+	ControllerInput ci = car.GetInput();
+
+	int marginLeft = 30;
+	int marginTop = 20;
+
+	int nameSpacing = 125;
+	int vecSpacing = 120;
+	int quatSpacing = 120;
+
+	int lineSpacing = 30;
+
+	cw.SetPosition({ x, y });
+	cw.SetColor(COLOR_PANEL);
+	cw.FillBox({ 420, 280 });
+	cw.SetColor(COLOR_TEXT);
+
+	this->drawStringAt(cw, "Player input", x + marginLeft, y + marginTop);
+
+	int currentLine = marginTop + 50;
+
+	this->drawStringAt(cw, "Throttle", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ci.Throttle), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+
+	this->drawStringAt(cw, "Steer", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ci.Steer), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+
+	this->drawStringAt(cw, "Pitch", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ci.Pitch), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+
+	this->drawStringAt(cw, "Yaw", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ci.Yaw), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+
+	this->drawStringAt(cw, "Roll", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ci.Roll), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+
+	this->drawStringAt(cw, "Handbrake", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ci.Handbrake), x + marginLeft + nameSpacing, y + currentLine);
+
+	// second column
+	currentLine = marginTop + 50;
+	marginLeft += 200;
+
+	this->drawStringAt(cw, "Jump", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ci.Jump), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+
+	this->drawStringAt(cw, "Jumped", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ci.Jumped), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+
+	this->drawStringAt(cw, "Activate boost", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ci.ActivateBoost), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+
+	this->drawStringAt(cw, "Holding boost", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ci.HoldingBoost), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+
+	this->drawStringAt(cw, "Dodge forward", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ci.DodgeForward), x + marginLeft + nameSpacing, y + currentLine);
+
+	currentLine += lineSpacing;
+
+	this->drawStringAt(cw, "Dodge strafe", x + marginLeft, y + currentLine);
+	this->drawStringAt(cw, sp::to_string(ci.DodgeStrafe), x + marginLeft + nameSpacing, y + currentLine);
+}
